requirements.txt

flask==2.1.3
pytorch==1.13.1
numpy==1.21.5
pyyaml==6.0
web3==5.29.0
gunicorn==20.1.0
